requirements.txt
langchain-groq
fastapi
python-dotenv
langchain_groq